requirements.txt
agate==1.9.1
annotated-types==0.7.0
asn1crypto==1.5.1
attrs==25.3.0
babel==2.17.0
backports.tarfile==1.2.0
boto3==1.40.35
botocore==1.40.35
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.0
daff==1.4.2
dbt-adapters==1.16.7
dbt-common==1.31.0
dbt-core==1.10.11
dbt-extractor==0.6.0
dbt-protos==1.0.375
dbt-semantic-interfaces==0.9.0
dbt-snowflake==1.10.2
deepdiff==8.6.1
dotenv==0.9.9
filelock==3.19.1
idna==3.10
importlib_metadata==8.7.0
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kafka-python==2.2.15
keyring==25.6.0
leather==0.4.0
MarkupSafe==3.0.2
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.1
networkx==3.5
orderly-set==5.5.0
packaging==25.0
parsedatetime==2.6
pathspec==0.12.1
platformdirs==4.4.0
protobuf==6.32.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
s3transfer==0.14.0
six==1.17.0
snowflake-connector-python==3.17.4
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
sqlparse==0.5.3
text-unidecode==1.3
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uv==0.8.19
zipp==3.23.0
